+++ crates/protocol-usb/CHANGELOG.md
@@ -59,4 +59,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 8 -->
+<!-- Increment to skip CHANGELOG.md test: 9 -->